Move resetting margin from areaSettingsPanel to marginService #804
