+++ visualization/app/codeCharta/state/store/dynamicSettings/margin/margin.service.spec.ts
@@ -2,14 +2,24 @@ import "../../../state.module"
 import { IRootScopeService } from "angular"
 import { StoreService } from "../../../store.service"
 import { getService, instantiateModule } from "../../../../../../mocks/ng.mockhelper"
-import { MarginAction, MarginActions } from "./margin.actions"
+import { MarginAction, MarginActions, setMargin } from "./margin.actions"
 import { MarginService } from "./margin.service"
-import { withMockedEventMethods } from "../../../../util/dataMocks"
+import { TEST_FILE_WITH_PATHS, withMockedEventMethods } from "../../../../util/dataMocks"
+import { CodeMapPreRenderService } from "../../../../ui/codeMap/codeMap.preRender.service"
+import { AreaMetricService } from "../areaMetric/areaMetric.service"
+import { DynamicMarginService } from "../../appSettings/dynamicMargin/dynamicMargin.service"
+import { setDynamicMargin } from "../../appSettings/dynamicMargin/dynamicMargin.actions"
+import { setAreaMetric } from "../areaMetric/areaMetric.actions"
+import { CodeMapNode } from "../../../../codeCharta.model"
+import _ from "lodash"
 
 describe("MarginService", () => {
 	let marginService: MarginService
-	let storeService: StoreService
 	let $rootScope: IRootScopeService
+	let storeService: StoreService
+	let codeMapPreRenderService: CodeMapPreRenderService
+
+	let map: CodeMapNode
 
 	beforeEach(() => {
 		restartSystem()
@@ -22,10 +32,19 @@ describe("MarginService", () => {
 
 		$rootScope = getService<IRootScopeService>("$rootScope")
 		storeService = getService<StoreService>("storeService")
+		codeMapPreRenderService = getService<CodeMapPreRenderService>("codeMapPreRenderService")
+
+		map = _.cloneDeep(TEST_FILE_WITH_PATHS.map)
 	}
 
 	function rebuildService() {
-		marginService = new MarginService($rootScope, storeService)
+		marginService = new MarginService($rootScope, storeService, codeMapPreRenderService)
+	}
+
+	function withMockedCodeMapPreRenderService() {
+		codeMapPreRenderService = marginService["codeMapPreRenderService"] = jest.fn().mockReturnValue({
+			getRenderMap: jest.fn().mockReturnValue(map)
+		})()
 	}
 
 	describe("constructor", () => {
@@ -35,6 +54,22 @@ describe("MarginService", () => {
 			rebuildService()
 
 			expect(StoreService.subscribe).toHaveBeenCalledWith($rootScope, marginService)
+		})
+
+		it("should subscribe to AreaMetricService", () => {
+			AreaMetricService.subscribe = jest.fn()
+
+			rebuildService()
+
+			expect(AreaMetricService.subscribe).toHaveBeenCalledWith($rootScope, marginService)
+		})
+
+		it("should subscribe to DynamicMarginService", () => {
+			DynamicMarginService.subscribe = jest.fn()
+
+			rebuildService()
+
+			expect(DynamicMarginService.subscribe).toHaveBeenCalledWith($rootScope, marginService)
 		})
 	})
 
@@ -55,6 +90,51 @@ describe("MarginService", () => {
 			marginService.onStoreChanged("ANOTHER_ACTION")
 
 			expect($rootScope.$broadcast).not.toHaveBeenCalled()
+		})
+	})
+
+	describe("reset", () => {
+		beforeEach(() => {
+			withMockedCodeMapPreRenderService()
+		})
+
+		it("should not call dispatch if dynamicMargin is false", () => {
+			storeService.dispatch(setDynamicMargin(false))
+			storeService.dispatch = jest.fn()
+
+			marginService.reset()
+
+			expect(storeService.dispatch).not.toHaveBeenCalled()
+		})
+
+		it("should set new calculated margin correctly", () => {
+			storeService.dispatch(setDynamicMargin(true))
+			storeService.dispatch(setAreaMetric("rloc"))
+
+			marginService.reset()
+
+			expect(storeService.getState().dynamicSettings.margin).toEqual(28)
+		})
+
+		it("should call dispatch after setting new margin", () => {
+			storeService.dispatch(setDynamicMargin(true))
+			storeService.dispatch(setAreaMetric("rloc"))
+			storeService.dispatch = jest.fn()
+
+			marginService.reset()
+
+			expect(storeService.dispatch).toHaveBeenCalled()
+		})
+
+		it("should not call applySettings if margin and new calculated margin are the same", () => {
+			storeService.dispatch(setMargin(28))
+			storeService.dispatch(setDynamicMargin(true))
+			storeService.dispatch(setAreaMetric("rloc"))
+			storeService.dispatch = jest.fn()
+
+			marginService.reset()
+
+			expect(storeService.dispatch).not.toHaveBeenCalled()
 		})
 	})
 })
