+++ visualization/app/codeCharta/ui/areaSettingsPanel/areaSettingsPanel.component.spec.ts
@@ -2,33 +2,24 @@ import "./areaSettingsPanel.module"
 import "../codeMap/codeMap.module"
 import "../../codeCharta.module"
 import { AreaSettingsPanelController } from "./areaSettingsPanel.component"
-import { TEST_FILE_WITH_PATHS } from "../../util/dataMocks"
 import { FileStateService } from "../../state/fileState.service"
 import { IRootScopeService } from "angular"
 import { getService, instantiateModule } from "../../../../mocks/ng.mockhelper"
-import { CodeMapNode } from "../../codeCharta.model"
-import { CodeMapPreRenderService } from "../codeMap/codeMap.preRender.service"
 import { StoreService } from "../../state/store.service"
 import { DynamicMarginService } from "../../state/store/appSettings/dynamicMargin/dynamicMargin.service"
 import { MarginService } from "../../state/store/dynamicSettings/margin/margin.service"
-import { setAreaMetric } from "../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
 import { setDynamicMargin } from "../../state/store/appSettings/dynamicMargin/dynamicMargin.actions"
-import _ from "lodash"
-import { AreaMetricService } from "../../state/store/dynamicSettings/areaMetric/areaMetric.service"
 
 describe("AreaSettingsPanelController", () => {
+	let areaSettingsPanelController: AreaSettingsPanelController
 	let $rootScope: IRootScopeService
 	let storeService: StoreService
-	let codeMapPreRenderService: CodeMapPreRenderService
-	let areaSettingsPanelController: AreaSettingsPanelController
 
-	let map: CodeMapNode
 	let SOME_EXTRA_TIME = 400
 
 	beforeEach(() => {
 		restartSystem()
 		rebuildController()
-		withMockedCodeMapPreRenderService()
 	})
 
 	function restartSystem() {
@@ -36,19 +27,10 @@ describe("AreaSettingsPanelController", () => {
 
 		$rootScope = getService<IRootScopeService>("$rootScope")
 		storeService = getService<StoreService>("storeService")
-		codeMapPreRenderService = getService<CodeMapPreRenderService>("codeMapPreRenderService")
-
-		map = _.cloneDeep(TEST_FILE_WITH_PATHS.map)
 	}
 
 	function rebuildController() {
-		areaSettingsPanelController = new AreaSettingsPanelController($rootScope, storeService, codeMapPreRenderService)
-	}
-
-	function withMockedCodeMapPreRenderService() {
-		codeMapPreRenderService = areaSettingsPanelController["codeMapPreRenderService"] = jest.fn().mockReturnValue({
-			getRenderMap: jest.fn().mockReturnValue(map)
-		})()
+		areaSettingsPanelController = new AreaSettingsPanelController($rootScope, storeService)
 	}
 
 	describe("constructor", () => {
@@ -68,28 +50,12 @@ describe("AreaSettingsPanelController", () => {
 			expect(MarginService.subscribe).toHaveBeenCalledWith($rootScope, areaSettingsPanelController)
 		})
 
-		it("should subscribe to CodeMapPreRenderService", () => {
-			CodeMapPreRenderService.subscribe = jest.fn()
-
-			rebuildController()
-
-			expect(CodeMapPreRenderService.subscribe).toHaveBeenCalledWith($rootScope, areaSettingsPanelController)
-		})
-
 		it("should subscribe to FileStateService", () => {
 			FileStateService.subscribe = jest.fn()
 
 			rebuildController()
 
 			expect(FileStateService.subscribe).toHaveBeenCalledWith($rootScope, areaSettingsPanelController)
-		})
-
-		it("should subscribe to AreaMetricService", () => {
-			AreaMetricService.subscribe = jest.fn()
-
-			rebuildController()
-
-			expect(AreaMetricService.subscribe).toHaveBeenCalledWith($rootScope, areaSettingsPanelController)
 		})
 	})
 
@@ -99,22 +65,6 @@ describe("AreaSettingsPanelController", () => {
 
 			expect(areaSettingsPanelController["_viewModel"].dynamicMargin).toBeTruthy()
 		})
-
-		it("should set given margin if dynamicMargin is false", () => {
-			areaSettingsPanelController["_viewModel"].margin = 48
-
-			areaSettingsPanelController.onDynamicMarginChanged(false)
-
-			expect(areaSettingsPanelController["_viewModel"].margin).toEqual(48)
-		})
-
-		it("should set new calculated margin if dynamicMargin is true", () => {
-			storeService.dispatch(setAreaMetric("rloc"))
-
-			areaSettingsPanelController.onDynamicMarginChanged(true)
-
-			expect(storeService.getState().dynamicSettings.margin).toEqual(28)
-		})
 	})
 
 	describe("onMarginChanged", () => {
@@ -122,57 +72,6 @@ describe("AreaSettingsPanelController", () => {
 			areaSettingsPanelController.onMarginChanged(28)
 
 			expect(areaSettingsPanelController["_viewModel"].margin).toBe(28)
-		})
-	})
-
-	describe("onAreaMetricChanged", () => {
-		it("should call potentiallyUpdateMargin", () => {
-			areaSettingsPanelController["potentiallyUpdateMargin"] = jest.fn()
-
-			areaSettingsPanelController.onAreaMetricChanged("SOME_METRIC")
-
-			expect(areaSettingsPanelController["potentiallyUpdateMargin"]).toHaveBeenCalled()
-		})
-	})
-
-	describe("onRenderFileChange", () => {
-		it("should not call dispatch if dynamicMargin is false", () => {
-			storeService.dispatch(setDynamicMargin(false))
-			storeService.dispatch = jest.fn()
-
-			areaSettingsPanelController.onRenderMapChanged(map)
-
-			expect(storeService.dispatch).not.toHaveBeenCalled()
-		})
-
-		it("should set new calculated margin correctly", () => {
-			areaSettingsPanelController["_viewModel"].dynamicMargin = true
-			storeService.dispatch(setAreaMetric("rloc"))
-
-			areaSettingsPanelController.onRenderMapChanged(map)
-
-			expect(storeService.getState().dynamicSettings.margin).toEqual(28)
-		})
-
-		it("should call dispatch after setting new margin", () => {
-			areaSettingsPanelController["_viewModel"].dynamicMargin = true
-			storeService.dispatch(setAreaMetric("rloc"))
-			storeService.dispatch = jest.fn()
-
-			areaSettingsPanelController.onRenderMapChanged(map)
-
-			expect(storeService.dispatch).toHaveBeenCalled()
-		})
-
-		it("should not call applySettings if margin and new calculated margin are the same", () => {
-			areaSettingsPanelController["_viewModel"].margin = 28
-			areaSettingsPanelController["_viewModel"].dynamicMargin = true
-			storeService.dispatch(setAreaMetric("rloc"))
-			storeService.dispatch = jest.fn()
-
-			areaSettingsPanelController.onRenderMapChanged(map)
-
-			expect(storeService.dispatch).not.toHaveBeenCalled()
 		})
 	})
