+++ visualization/app/codeCharta/ui/areaSettingsPanel/areaSettingsPanel.component.ts
@@ -1,8 +1,6 @@
 import "./areaSettingsPanel.component.scss"
 import { IRootScopeService } from "angular"
-import { CodeMapNode, FileState } from "../../codeCharta.model"
-import { hierarchy, HierarchyNode } from "d3-hierarchy"
-import { CodeMapPreRenderService, CodeMapPreRenderServiceSubscriber } from "../codeMap/codeMap.preRender.service"
+import { FileState } from "../../codeCharta.model"
 import { FileStateService, FileStateServiceSubscriber } from "../../state/fileState.service"
 import { StoreService } from "../../state/store.service"
 import { setDynamicMargin } from "../../state/store/appSettings/dynamicMargin/dynamicMargin.actions"
@@ -10,18 +8,8 @@ import { setMargin } from "../../state/store/dynamicSettings/margin/margin.actio
 import _ from "lodash"
 import { DynamicMarginService, DynamicMarginSubscriber } from "../../state/store/appSettings/dynamicMargin/dynamicMargin.service"
 import { MarginService, MarginSubscriber } from "../../state/store/dynamicSettings/margin/margin.service"
-import { AreaMetricService, AreaMetricSubscriber } from "../../state/store/dynamicSettings/areaMetric/areaMetric.service"
 
-export class AreaSettingsPanelController
-	implements
-		CodeMapPreRenderServiceSubscriber,
-		FileStateServiceSubscriber,
-		DynamicMarginSubscriber,
-		MarginSubscriber,
-		AreaMetricSubscriber {
-	private static MIN_MARGIN = 15
-	private static MAX_MARGIN = 100
-	private static MARGIN_FACTOR = 4
+export class AreaSettingsPanelController implements FileStateServiceSubscriber, DynamicMarginSubscriber, MarginSubscriber {
 	private static DEBOUNCE_TIME = 400
 	private readonly applyDebouncedMargin: () => void
 
@@ -34,16 +22,10 @@ export class AreaSettingsPanelController
 	}
 
 	/* @ngInject */
-	constructor(
-		private $rootScope: IRootScopeService,
-		private storeService: StoreService,
-		private codeMapPreRenderService: CodeMapPreRenderService
-	) {
+	constructor(private $rootScope: IRootScopeService, private storeService: StoreService) {
 		DynamicMarginService.subscribe(this.$rootScope, this)
 		MarginService.subscribe(this.$rootScope, this)
-		CodeMapPreRenderService.subscribe(this.$rootScope, this)
 		FileStateService.subscribe(this.$rootScope, this)
-		AreaMetricService.subscribe(this.$rootScope, this)
 
 		this.applyDebouncedMargin = _.debounce(() => {
 			this.storeService.dispatch(setMargin(this._viewModel.margin))
@@ -52,15 +34,10 @@ export class AreaSettingsPanelController
 
 	public onDynamicMarginChanged(dynamicMargin: boolean) {
 		this._viewModel.dynamicMargin = dynamicMargin
-		this.potentiallyUpdateMargin()
 	}
 
 	public onMarginChanged(margin: number) {
 		this._viewModel.margin = margin
-	}
-
-	public onRenderMapChanged(map: CodeMapNode) {
-		this.potentiallyUpdateMargin(map)
 	}
 
 	public onFileSelectionStatesChanged(fileStates: FileState[]) {
@@ -77,35 +54,6 @@ export class AreaSettingsPanelController
 	public onChangeMarginSlider() {
 		this.applyDebouncedMargin()
 		this.storeService.dispatch(setDynamicMargin(false))
-	}
-
-	public onAreaMetricChanged(areaMetric: string) {
-		this.potentiallyUpdateMargin()
-	}
-
-	private potentiallyUpdateMargin(map: CodeMapNode = this.codeMapPreRenderService.getRenderMap()) {
-		if (this._viewModel.dynamicMargin && this.storeService.getState().dynamicSettings.areaMetric && map) {
-			const newMargin = this.computeMargin(this.storeService.getState().dynamicSettings.areaMetric, map)
-			if (this._viewModel.margin !== newMargin) {
-				this.storeService.dispatch(setMargin(newMargin))
-			}
-		}
-	}
-
-	private computeMargin(areaMetric: string, map: CodeMapNode): number {
-		let leaves = hierarchy<CodeMapNode>(map).leaves()
-		let numberOfBuildings = 0
-		let totalArea = 0
-
-		leaves.forEach((node: HierarchyNode<CodeMapNode>) => {
-			numberOfBuildings++
-			if (node.data.attributes && node.data.attributes[areaMetric]) {
-				totalArea += node.data.attributes[areaMetric]
-			}
-		})
-
-		let margin: number = AreaSettingsPanelController.MARGIN_FACTOR * Math.round(Math.sqrt(totalArea / numberOfBuildings))
-		return Math.min(AreaSettingsPanelController.MAX_MARGIN, Math.max(AreaSettingsPanelController.MIN_MARGIN, margin))
 	}
 }
